+++ examples/main/modules.tf
@@ -1,0 +1,6 @@
+module "azure_region" {
+  source  = "claranet/regions/azurerm"
+  version = "x.x.x"
+
+  azure_region = "eu-west"
+}